requirements.txt
openai 
python-dotenv
datetime
requests
pydantic